requirements.txt
Django==3.2.9
python-decouple==3.5
django-taggit==1.5.1
Pillow==8.4.0
